Rename file multiple-column.vue to MultipleColumn.vue
